Remove support for Tuskar
This change removes Tuskar support from tripleo-common. Currently the only
consumer of tripleo-common is the python-tripleoclient which is removing
Tuskar support in the dependant change ID below.

Future work in the client will be with Heat directly and maintaining
support for Tuskar in tripleo-common just adds a large maintenance burden
since it won't be used.

Depends-On: Ic6fd4568054fe91b8fc898728022a7d7ed8cf3cf
Change-Id: If8c2ab70edb5325bde2524a9a5d824605cadcc18