Add tripleo.sh
tripleo.sh is a script to automate a TripleO setup. It's goals are to be
used in aiding:

- developer setups
- CI
- documentation generation (hopefully)

It's not a new CLI, or end user facing wrapper around existing TripleO
CLI's.

tripleo.sh should never contain any "business" logic in it that is
necessary for a successful deployment. It should instead just mirror the
steps that we document for TripleO end users.

Change-Id: I9396a487964955f59584965ecbfd4c88cca87c0e