Adds a simple overcloud tenant vm ping test to tripleo.sh
As discussed at [1] this adds a simple overcloud tenant vm startup,
with a floating ip attached, which is then pinged. The goal is to
eventually add tempest support though consensus seems to be we should
carry at least a ping test.

This has now been re-implemented with a heat template, which we
carry here for now (but can be moved to a better place if necessary)

The aim here is to add this to tripleo.sh to it primarily will be
consumed by toci_instack for ci

[1] https://trello.com/c/OQlIbGoK/19-toci-start-a-user-instance-and-ping-it

Change-Id: Ifbbaf0a84dd9e0530fa8b9a42a0f66793e0df94d