Add scripts/upload-swift-artifacts
This patch adds a new script/tool that can be used to upload
a set of file(s) into Swift. The tool automatically creates and
configures a Swift Temp URL in the undercloud and then generates
the required heat environment in $HOME/.tripleo/environments/ so that it
will automatically be used by heat deployments.

Change-Id: If63260e59ae44b172a76222d591d4e62d9432655
Implements: blueprint puppet-modules-deployment-via-swift