Make coverage tests work
The default package name for coverage doesn't work, so we need to
explicitly state it in tox.ini. We should also exclude the tests
directory as that code will typically be 100% tested, and if it's
not we don't really care.  Also excludes the coverage results from
git so it doesn't show up in git status.

Change-Id: I9e99b2c7c215a5de3c76cb2daf0f468f445774c3