Add OVERCLOUD_IMAGES_ARGS environment variable
Add a new OVERCLOUD_IMAGES_ARGS environment varible to allow modify the
parameters which we call the `openstack overcloud image build`. For
backwards compatibility and general purpose, this variable is
initialized to `--all`.

This change opens the door to build a single image by overriding the env
var to values like `--type overcloud-full` or more complex calls.

Change-Id: Iaa4d20f8cfdbd92744186e54cf35a21962ef6f6e