Convenience methods in base model class
This change adds a BaseModel type which Plan inherits from along with
implementations of __eq__ and __repr__ to assist with testing and
debugging.

The __repr__ is a compact yaml representation which starts with the
class name.

The __eq__ asserts the other is of the same type, and then compares the
class __dict__ (which includes the class attributes)

Change-Id: I3a6aafc680637b2ec379f206a731796c24d5c7bb