Update pingtest for newer heatclient
I assume this works in tripleoci now because we are
pinned to an old delorean repo. Using current, the
`heat output-show` command returns a table which is
not directly pingable. The '-v' flag makes it return
a value instead.

We will definitely need this patch to bump the
current-tripleo delorean hash.

Change-Id: I25e46a3d7ea10cd1748a9c602661f4d42f2b478a