Use Fedora image for ping test
It appears the Cirros image doesn't have no_timer_check, which
sometimes causes the image to hang on boot.  Let's use a fedora
image that we know does have no_timer_check.  This will cause the
test to take a little longer because it doesn't boot as fast as
Cirros, but that's way better than waiting two hours for a recheck.

I'm using the old f21 image because that's what we were using in
the past for instack CI, and it seemed to be reliable back then.
Once we've got this stabilized a bit we can look at moving to a
new fedora image (although for some reason the f21 image is quite
a bit smaller than more recent ones, so we might just want to keep
it).

Change-Id: Iaf2257a76a8214f165ff412574ea77ca5d07cb77