Extend ping test create timeout
Now that we're using a Fedora image, it appears to take longer to
spawn the VM on some jobs.  One that failed took 167 seconds just
to start the instance, which is uncomfortably close to our timeout
of 180 seconds.  Other than the long spawn time everything appeared
to have worked correctly, so it probably would have passed otherwise.

Bumping the timeout to 360 to match the ping timeout.  In
successful runs it generally won't take that long anyway so it
shouldn't add any significant time to CI runs.

Change-Id: Icd587c8858bbf817c4405932e246a4b890ded11f