Move delorean pin
We now have fixes for everything except the ones
dependant on unpinning puppet-nova. This updates the
delorean pin to a specific delorean repo which has
passed for me locally with both HA and nonHA setups.

If this passes, we can merge it with the dependant
patches, and then set it back to the current-tripleo
symlink once that is moved forward.

Depends-On: I8b6d907fe46579007f33a7e69e7d67e79582426c
Depends-On: Ifb8c23c81c665c2732fa5cd757760668b06a449a

Change-Id: I47b4470070ec42a108a90cbae6250ed279cd257d