Use m1.small for ping test
m1.tiny does not have a large enough disk for the fedora image, but
m1.small does.  The only downside is that it is a 2 GB ram image.
However, our compute nodes now have 5 GB of RAM so there's plenty
available, and with such a short-lived VM it's likely that the
actual difference in memory used between the two is minimal.

This will allow us to remove the m1.demo flavor creation from the
overcloud deploy since it was only ever there for test purposes.

Change-Id: I6b1e35c00e51ae9da1830a1e6d714e5f0858bcdb