Add upload-puppet-modules script
This utility script can be used to tarball the provided directory
and upload it via upload-swift-artifacts. A tar transformation
is performed so that the end user can specify the actual directory
of puppet modules rather than having to nest it within an
etc/puppet/modules directory so it gets installed correctly on the
remote server.

Change-Id: Ie611e2b866549bb2401c4a684e89cca48264680f