Prevent replacement of Port, Net and Subnet
There have been cases reported of Neutron resources being reallocated on
some stack-updates. (E.g. the ports which are used for VIP allocation
and feeding VIP info to overcloud config files.) This is quite similar
in the OS::Nova::Server replacement issue in the sense that TripleO
clouds currently cannot recover from this happening. While the triggers
of this behavior might have external causes due to manipulation with
templates or environment files (we haven't found any root cause in
TripleO code yet), in general this is an issue that we should probably
prevent if we can.

Closes-Bug: #1572609
Change-Id: I1bfd3b4c2bf23ddb96905ddc654b807955668922