Import initial_state logic from tripleoclient
Ironic API v1.11 introduced a new ENROLL state for newly-registered
baremetal nodes. Logic to allow the user to specify which initial state
to use (either ENROLL or the older AVAILABLE) was previously in
python-tripleoclient; this patch moves it into this repo, along with
supporting utility functions and tests.

Specifically, we

 - add a `provide` keyword argument to utils.nodes.register_all_nodes;
   when set to True, newly-registered nodes should be transitioned to
   the AVAILABLE state;
 - import the utility functions wait_for_provision_state and
   set_nodes_state from python-tripleoclient;
 - change set_nodes_state to return a list of node objects rather than
   generating a sequence of UUIDs. This makes its results more useful to
   client code.

Change-Id: I34d48293c90f9b3b18d05272166a72b61b209b66