Allow updating of nodes in baremetal import
Right now, if you attempt to update a node's details using the
baremetal import command and the node has both a mac and a pm_addr
specified, it will fail because _get_node_id finds both the mac
and pm_addr match and adds the id to a list twice.  Since it's
perfectly fine for a node to match both of those fields, we
shouldn't error in that case.  Using a set instead of a list for
the data structure makes this work as expected.

Change-Id: I4b9fb1531c59a090b67f19d673c21c22b4f9523b
Closes-Bug: 1586523