Minor update to the upgrade-non-controller.sh
Minor fixup for the --query function which was intended to provide
some insight into whether the node was upgraded or not. Since there
is no easy way to tell, just check if the upgrade script is on the
node yet and check what packages if any were recently updated. This
should at least set expectations wrt what --query can do. Also adds
short opts on request -u -s -q

Part of updating the upgrades workflow for M..N

Implements: overcloud-upgrades-workflow-mitaka-to-newton
Change-Id: I16b87d34070221d2f37df9ddf99922c44a9f0364