Disable retry files for ansible validations
By default on failed runs, ansible will try to write a .retry file in
the same directory where the playbook is located. In the case of
tripleo-validations this is a read only directory under /usr/share.

This commit disables retry files until we determine they can be useful
to us and find a writable place to store them.

Change-Id: Ied2a550930dbf396a06c91fa00c2d066e8b39562