Adds action for template processing
This patch moves the code that deals with calling out to
heatclient to process multiple files and environments from
utils/templates.py into an action that will provide a base object
for the deployment and parameters actions to inherit from. This
patch also removes the resulting dead code (utils/templates.py)
and its associated test.

Co-Authored-By: Dougal Matthews <dougal@redhat.com>
Change-Id: I8efb53ac14c3b952743d1871df8d63d32685b6b1