Add support to image build yaml input to handle env vars
The inputs to DIB are (in practice) a combination of parameters
and env variables.  This allows the extra_options dict to be used
to pass env variables, and support (if present) setting those
variables in the enviornment.  This does not change the existing
support to handle ELEMENTS_PATH.  variables are set as a top
level dict variable 'environment:' in the input yaml.

Change-Id: Iad2cb67b73e0d0a05b84d045e8147349ad5006e2