Fix upload-swift-artifacts work new OSclient
This resolves a regression in 7d41589d59d42aa4f597d3787498130d1fdf5c43
where we again broke this shell script with newer version of
openstackclient. In that commit we added an
'openstack endpoint show swift' which again causes errors with
the latest openstackclient.

This patch reverses the logic so that we try to calculate the
swift internal URL with the new command first. This won't error
on the old version of openstackclient so everything should be happy
I think.

Change-Id: If46687dfb0f2ad2a116eae41422ea160f61026cd
Closes-bug: #1618488