Clearer error when the Mistral env already exists
When creating a plan with custom templates, if the Mistral environment
already exists the error message returned is unclear: "Duplicate entry
for Environment: ['name']". This changes the message to be the same
than when creating a plan using the default templates: "The Mistral
environment already exists".

Also, to avoid copy-pasting the mocks for each new unit test, I
refactored them so that the mocks are all set up once in one place.

Change-Id: Ic5855ff553c54a1b9842602812e4f65c49989ae1
Closes-Bug: #1620747