Port password generation from tripleoclient to tripleo-common
At the moment the CLI generates various passwords, this logic is
then not available to the GUI. We should move this to further
standardise the process.

This patch ports the necessary utility methods and adds an action
to generate and store passwords in a mistral environment or retrieve
previously generated passwords from the mistral environment.  The action
will generate and replace any missing passwords.  The passwords are
stored in parameter_defaults and are used in any subsequent calls to heat
(e.g. Parameters and Stack Creation).

The order of the merge of parameter_defaults and passwords allows
a user to override a password value.

The action is added to workflows in the plan_management workbook.

Closes-Bug: #1621097
Change-Id: Ic476a09f7981d4e6ee12e05b333a18cda5b4626b