Support node untagging
The tag_nodes workflow now takes two lists of nodes: nodes to tag, and
nodes to clear, and of course a role.  First, the workflow tags nodes
with a role, then it clears the other node tags, and finally it updates
the role count.

Calling update_node_capability with a value of None will clear the tag
from the node.

Change-Id: I1e2988297d56fb4726cd3cae79e79dda889471e1
Closes-bug: #1625264
Co-Authored-By: Honza Pokorny <hpokorny@redhat.com>