Fix upload-swift-artifacts for old OSclient
If one has the old openstackclient, SWIFT_INTERNAL_URL
never gets set in upload-swift-artifacts because
'grep internal' fails in the initial attempt pipeline
and the script immediately exits at that point.

Modify the initial attempt to set SWIFT_INTERNAL_URL to
return success even if the pipeline of commands has an
error.

Change-Id: I0cdb3be1d8e984e271cc940be053117ae503ca81
Closes-bug: 1628065