Only use parameter_defaults for scaling down
We were using parameters for scaling down but parameter_defaults for
scaling up. The old parameters from scale-down still persisted in the
stack and took precedence over the scaled-up parameter_defaults, so the
scaled-up values didn't propagate.

We already update parameter_defaults in the deployment plan in that same
method, so all we need to do is remove the 'parameters' part.

Change-Id: I90a195a66b15c7c9057bfb415b6a730bbd9a939c
Closes-Bug: #1641142