Remove python-networking-cisco from overcloud-full image
The networking-cisco library is broken on master due to deprecated
code being removed from neutron. I attempted a fix[1], but CI on
that patch shows there are other places that will break even after
that fix. We can not block promotion of TripleO CI due to broken
vendor plugin, so this patch removes the package from the overcloud
image. When the library is updated to work with master neutron, we
can revert this patch.

[1] https://review.openstack.org/#/c/397239/

Change-Id: I4c88112213408a5111be0e0bd2f90590ddc30887
Related-Bug: 1641311