Revert changes to CephClusterFSID generation
This patch reverts the two commits listed below. There was a problem
with CephClusterFSID generation as it was created by the password
generation code in Mistral in Newton. However, this password generation
didn't work for upgrades as it didn't take into account upgrades and
passwords that were in use in an existing Heat stack.

This issue was resolved in I3ea6bbd0d9c5dd345b8a4a26a1788326e09d4209.
Now when a deployment plan is created, if there is an existing Heat
stack it takes the parameters from the Heat stack and uses those, rather
than regenerating. This change should also resolve the issue where the
CephClusterFSID was regenerated.

Revert "Generate CephClusterFSID for new stacks"
This reverts commit 20167e850a6bcb1fbb1bfa8cb030000600c3dc27.

Revert "Revert "Add CephClusterFSID to generated passwords""
This reverts commit ad64050485f79d2330c68b76a3f21b4c0d9a3e33.

Related-Bug: #1636555
Change-Id: I10b5613eda4bd47554a4f5e9f57218010b835fe7