Add utility workflow to wait for stack COMPLETE or FAILED
In an attempt to move blocking on the stack moving back to a COMPLETE
or FAILED state out of the client and into the workflows, here
is a utility workflow to do just that. First user will be the node
delete workflow. Also add a workflow for waiting for a stack to go to
IN_PROGRESS.

Change-Id: Ibbb359de8a873ee76b0e274e631307d4e4804fca