Remove remaining vendor plugins from default image YAML
Trunk is again broken by a vendor plugin not keeping up with
deprecations in neutron. This time it is the bigswitch plugin
that is using code from neutron that was moved to neutron-lib.

There is not really any reason to have these vendor plugins
installed in the default YAML. If the vendor wants to build images
for TripleO in their own CI, it is simply a matter of passing an
extra yaml with the packages key.

Change-Id: Ib7c10f2dbd692494bd1990c3a2988308096743e7
Closes-Bug: 1649284