Set imageext for IPA ramdisk
Without this, the existing image logic looks for a file named
ironic-python-agent.qcow2, which will never exist since IPA is a
ramdisk.  Note that there is no initrd type in dib, so we have to
leave that as qcow2 even though it's wrong.  I left a comment in
the yaml explaining why.

Additionally, this required some modification of the logic around
the image name.  For the purposes of checking existence, we need
the extension.  For the diskimage-builder call, we do not.  DIB
strips off any extension that matches the image type anyway, but
this does not work properly if we pass a type of qcow2 with an
extension of initramfs, so we end up with a file named
ironic-python-agent.initramfs.initramfs.

Change-Id: Ic5757c1f363979a6c52830bfa6c42c2b57740fd5
Closes-Bug: 1654002