Use latest containers for master branch
The master branch should use containers tagged with 'latest' instead of
a named openstack version.

Also stop pulling the containers that are not currently being used.
With the new arch that is in the pipes, we use the host's
os-collect-config agent and also use the target containers directly to
generate the configuration files with puppet, making the heat-agents
container useless. The master branch also stopped using the ovs
containers in favor of the host's ovs.

Finally, remove the push_destination attribute to make it default to
the host's control plane IP address on port 8787.

Change-Id: Ifc0939d08ac9a9d35745c81a6a5f719ccee02b04
Depends-On: Ibcff99f03e6751fbf3197adefd5d344178b71fc2