Use the builtin Mistral engine command 'fail' when failing workflows
In changes 12f7345, c2fb9b0, 6051a74, c7b01eb and ab946c4 Mistral
Workflows were marked as failed if they sent a Zaqar message with the
status "FAILED". This is correct, but it turns out there is an easier
way to do it.

Mistral has a engine command designed for this purpose, using it removes
some of the duplication we have incurred.

We use on-success rather than on-complete to surface the best error on
the workflow. If the Zaqar queue post fails, then the workflow will fail
with this error. If, on-complete we then manually mark the workflow as
failed it will ovrwrite the Zaqar error with a generic one.

Closes-Bug: #1664918
Related-Bug: #1651704
Change-Id: I9ef9cfef1b8740a535e005769ec0c3ad67ecb103