Re-enable tee behavior for image builds
The previous client-based image builds tee'd the output to both a
log file and stdout.  With the yaml-based image builds we lost that
behavior, which makes for a bad user experience when the console
just seems to sit there for 30 minutes while images build, with no
feedback on what's going on.  In addition, due to the default
openstackclient logging configuration, we don't even see the info
message telling the user which file contains the log output.

This change uses the same method as the instack-undercloud
_run_live_command method to tee the output to both the Python
logger and the log file.  It also ensures that our local logger is
configured such that info level messages will be seen.

Change-Id: I7e17a18a8bc170ea70955c87a84ca401cb44aa19
Closes-Bug: 1653768
Closes-Bug: 1646226