Update to a newer hacking library
The older hacking library has a cap on pbr <2.0, with the recent 2.0.0
release of PBR it's causing failures in the pep8 job.

hacking isn't kept in sync via the typical proposal-bot updates.

Do it manually to clear the gate issue.

For newton branch we need to set 0.11.0 as minimal version to make
it compatible with global-requirements in stable/newton.

Change-Id: Icb4ad6380209fe006edeb26cfce31553cb1eb555
Related-Bug: #1668848
(cherry picked from commit 7f0902ba91573d66f0282414cfb86caee4b7bc79)