Fix using old Swift rings when creating a new stack
The container to store Swift rings is not deleted when the stack is
deleted. This results in re-deploying old rings when deploying a new
stack.

This patch creates a copy of the old rings with a timestamp suffix and
removes the original object. Doing this will create new rings when
creating a new stack.

Closes-Bug: 1671783
Change-Id: Ie6363987ada9b516064c1ed3b215f809c3924393