Library wrapper around kolla-build tool
This adds a python library which allows container images to be created
using the kolla-build tool. The yaml format used by
contrib/overcloud_containers.yaml is used to specify which images to
build.

It is also possible to specify the path to kolla conf files.
Every kolla option can be set by a conf file, so this is sufficient to
modify kolla-build behaviour.

The namespace and the tag in overcloud_containers.yaml is ignored
since this is driven by the kolla configuration.

The image prefixes like 'centos-binary-' are removed before passing
the remaining image name to kolla as a regex match.

This will be used by the new command "overcloud container image build"

Change-Id: I061f626fdb3d71613aa23436bf6c53cf4de62213