Add an action to fetch and flatten the heat resource tree and parameters
tripleo.parameters.get action provides the heat tree and the
parameters as nested stack. In order to easily consume it, the
nested stack should be flattened. Added a new action which will
flatten it as it is done in tripleo-ui, so that it can be common
between ui and cli.
Implements: blueprint tripleo-derive-parameters

Change-Id: I8931a1c2daef53180bb97a58f4cfb7a257496060