Remove Percona-Server-shared-56 from openstack-base image
Since Kolla commit d2c8372e99, yum is configured to fail the docker
build if a specified package is missing. This causes our image build to
fail on openstack-base because we disable external repos and yum can
not find a package for Percona-Server-shared-56. This commit blacklists
it from the openstack-base image.

Change-Id: Iba8d018b36965b49535a743fcd34428b72f3b561