Run introspection 20 nodes at a time
Some larger deployments can run into issues when introspecting a large
number of nodes. This patch updates the introspect workflow to run
introspection up to 20 nodes at one time.

The choice of 20 nodes per-batch is based on this comment:
https://bugs.launchpad.net/tripleo/+bug/1672854/comments/10

This patch moves introspection into a workflow that starts and waits for
it to finish. This means we can then simply run this workflow with
"with-items" and use Mistrals native support for concurrency limiting.

This tripleoclient patch allows users to change the number of nodes that
are introspected in parallel. https://review.openstack.org/#/c/451228/

Closes-Bug: #1669830
Change-Id: I2cb0a9247694c27782cce71b74feb5ac3ad4ddec