Adding the option for the stack id to stack delete
Right now it looks for stack_name only but tripleoclient is sending
the stack_id but this workbook looked only for stack_name so the stack was
never found and assumed deleted.

Because of that both openstack stack delete <stack> and openstack overcloud
delete <stack> would return instantly and wait for the stack to be
deleted.

Change-Id: I111e04696e945088aa626c1bb71582bd72390b8e
Closes-Bug: #1663423