Install os-net-config as a package, not an element
This change drops use of the os-net-config element from
t-h-t and simply installs the tool as a standard package.

The use of os-net-config has been deprecated in favor of
the new template styles for awhile now. The new hook gives better
signals when errors occur configuring the physical networks.

Change-Id: Ie4f5d771a16ea453b470be8ea103b2bde4aa892a
Related-bug: #1695091