Write action to get private key from the validations keypair
Currently there's only the action to get the public key, however
it's useful to get the private key as well. For instance, in cases
where we're using an ansible playbook. This way, we can use the
existing utilities to copy the public keys to the nodes, and then
use the private key for running ansible.

Change-Id: Ia44651d3d5908fa7ecd7e7ea2e288f4fe9faa7cf