Bump the default bare metal (ironic) API version in the overcloud to 1.34
This is the latest Pike API version, and it allows access to all new features
to the users of the client. No breaking changes were made to the API since Ocata.

Also issue a deprecation release note to inform the users that ironicclient will
change the default API version to latest, and we will follow this change for
the overcloud.

Depends-On: I3f8fbd38bb81930f8b5771faaef6005e689f4f26
Closes-Bug: #1706408
Change-Id: I4a5787effd3a3e62e6c16748f10471e3b0f6360f