Default to host CPU architecture if none is supplied
Currently the only way to specify the CPU architecture of the overcloud
images is by setting it in the appropriate yaml file.  This is fine
until you need to install a single package on multiple CPU architectures.
Then you're either going to attempt to build the wrong image type or you
have to edit the yaml files, which makes problems for upgrades.

This change removes the default architecture from the yaml files and
defaults to the appropriate DiB architecture based on running kernel.

This enables builds on multiple CPU architectures from a single
unmodified package

Change-Id: I34f00a30c1e07869243fb95a437e05d6973a97da