Parameterize Ansible skip_tags
Make Ansible's skip_tags a parameter to the Mistral workflow.
If using an image based deploy, ansible_skip_tags will default
to 'package-install,with_pkg'; this value causes ceph-ansible
to not install any packages, which we want for imaged deploys
because the image already has the package. If using already
deployed servers, pass '' to ansible_skip_tags (or any other
tag except 'package-install,with_pkg'), so that the package
installation tasks in ceph-ansible are not skipped.

Change-Id: Ia42cf104487193489499e152cc38cb7eabb9d567