Remove svg links
These links are breaking the docs build for tripleo-common on
stable/ocata effectively blocking us from merging any code on that
branch.

The bug was reported to sphinx and has been fixed:
https://github.com/sphinx-doc/sphinx/issues/3212

However, we don't have the newer version of sphinx in stable/ocata due
to the pin in test-requirements.txt coming from the global requirements
repo.

Change-Id: I06da2e01b898a303c445331387850a8cf1e24827
Partial-Bug: #1709152