Upload remove default for pull_source
Currently the pull_source defaults to "docker.io". This is unnecessary
as docker.io is the default registry anyway when no registry is
specified in the namespace.

Also having this default prevents fully qualified image paths being
specified in the imagename. This is desirable now that we use full
paths in the heat parameters.

This also allows the prepare command to deprecate the --pull-source
argument and document having full image paths in the --namespace
argument.

Change-Id: I1769c7cf1c058328a8490228bdc4719e8d216ccb
Partial-Bug: #1708967
(cherry picked from commit c58ac826df6628629eeddf9e9a5b2f2be5578191)