Add clustercheck healthcheck
Let's add a healthcheck for the clustercheck container.
We look for host and port in the /etc/xinetd.d/galera-monitor file
and if they should not be set or present we use port 9200 and the
hostname itself as default.

Since the clustercheck container uses the very same mariadb container
image, we test for xinetd's process and configuration file to decide
if we are using the container image for clustercheck or for mariadb.

Change-Id: I46be211323442d309d7bb000bede02b0e8ccb512
(cherry picked from commit 08d6641218b2933d8f0a0f645b96b08ef70e0b71)