Added a kolla config file for building contaner images with qemu uid
Guest VMs created for DPDK will have the vhost-user socket files
created by qemu and used by openvswitch. The uid of qemu in host
is 107 where as the uid set by kolla config is 42427. Because of
this mismatch, VM creation is failing because of permission.

Setting the same uid and gid for qemu user for building kolla
images to ensure that the files are created and used with same
permission. This patch adds a config file for TripleO to be
used for building kolla images, which will have the qemu
user id values override to match the host values.

Change-Id: I9a2d6b48eb44e294a8501ac0768ea0b97abb4b5c
Partial-Bug: #1712732
(cherry picked from commit fcdca87024d99e3eac61bdb227f179b286ca6e04)