Don't mock mkdtemp
As far as I can tell there's no reason to do this, and it breaks
the ability for NestedTempfile to clean up after the tests by
forcing the file to go to the system /tmp directory (which I'm
pretty sure is a security issue too...).  It's also causing some
weird errors in the gate so let's remove it.

Change-Id: I04f3f569ff904e002758098cfd90b5bbb31c6afc
Closes-Bug: 1714306
(cherry picked from commit 0be202dcd102ec06c78059472912938175cc5740)