Remove healthcheck command.
We want to leave the healtchecks in place in the containers, but not
set the command.  Since we reuse the containers for different purposes
in tripleo, the healtchecks run even if the intended service is not.
eg during configuration of containers, updating etc.  This change
removes the healtcheck command, which will be added to each service in
the service templates.

Change-Id: Ia2889d5c6a24de6e62e60deb3fa719fc437cbcb6