Fix chown command in sudoers file
The tripleo-common sudoers file allows an owner change on the
validations identity file. The command is only allowed with the `-h`
option, but this option isn't used everywhere in the code. This breaks
validation runs started through Mistral.

This patch changes the tripleo-common sudoers file to allow the chmod
command with and without the `-h` option.

Change-Id: I8770fc356a46a7a13ba3823a5d157104d08c087a
Closes-Bug: #1716625