Add logging to capabilities filter
The capabilities filter gets incorrectly blamed for a lot of
deployment failures because on a retry of a node deployment the
filter has to fail because there is only one node that can match
when using predictable placement.  However, we don't have any
logging to help determine why the filter fails.  This adds logging
to explain why nodes fell through, which will hopefully help with
debugging these problems.

Change-Id: I702209e9b5a1e546bf8cb784eec8bdc3c97c63eb
Closes-Bug: 1718502