Always finish by sending a Zaqar message on error
Add a task default to send a Zaqar message on an unhandled error. This
is useful in this workflow due to the complexity and as we have seen
some errors that cause the workflow to fail but the CLI and GUI never
recieve notice of this.

Partial-Bug: #1716671
Change-Id: I84a0fb8d039e9f1b6b124ee003c5cfb98b7f30db
(cherry picked from commit 064ae01d31440d458b40678aae6685aeffc5e127)