Don't wait for stack in progress in delete
We don't need to wait for the stack to be in progress when deleting it,
and it can in fact be a race condition if it disappears too fast. Just
skip to check if it still exists.

Change-Id: Idbbb7afd4ab0157c870e1c1e3cd1666fb99651cc
(cherry picked from commit 9178796142da747cf23d72c7a0b6b7c9554e90f4)