Adds detection for legacy API network resource
Currently in network_data.yaml there is compat_name which is used as a
workaround for upgrades with regards to a bug where the Heat resource
for API network was being created as InternalNetwork instead of
InternalApiNetwork.  This workaround does work for upgrades, but
consequently also causes all future new deployments to have the wrong
resource name.

This patch adds detection, so that if the legacy resource is detected
already in the overcloud stack, then the compatibility will be enabled
if not already set in network_data.yaml.  This allows for the removal of
compat_name from the network in network_data.yaml and thus removing the
incorrect name for future deployments.

Partial-Bug: 1718764

Change-Id: I695259ad87e2303f948875078bccb619786470e0
Signed-off-by: Tim Rozet <trozet@redhat.com>