Get a static inventory for upgrade-non-controller ansible runs
This is mainly used together with the newly added --overcloud-user
option and allows us to specify a custom overcloud admin user
e.g. in split stack environments 'stack' vs 'heat-admin'. The
static inventory is generated with the specified user.

As a further optimisation this also adds a new --inventory (-I)
option with which a previously generated inventory can be
specified.

This is also expected to be more efficient since we now only
generate the inventory at most once per node/invocation.

Change-Id: If0d5ca461ce6faf84c26a75bf9ab9e72bb061e7a
Closes-Bug: 1724552